Componentize "home-info" section
