+++ vecep/src/views/contents/home/styles.js
@@ -1,6 +1,4 @@
 import styled from 'styled-components';
-import { withStyles } from '@material-ui/core/styles';
-import { Button } from '@material-ui/core';
 
 const HomeContainer = styled.div`
   display: flex;
@@ -22,48 +20,8 @@ const HomeContainer = styled.div`
 
   & > div {
     box-sizing: border-box;
-    padding: 0 100px;
-  }
-
-  .home-intro {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding-top: 70px;
-    padding-bottom: 50px;
-
-    .text {
-      font-size: 1.5em;
-
-      span {
-        display: block;
-        font-size: 1.2em;
-        margin-block-end: 0.83em;
-        margin-inline-start: 0px;
-        margin-inline-end: 0px;
-        font-weight: bold;
-      }
-
-      @media only screen and (max-width: 1100px) {
-        width: 100%;
-      }
-    }
-
-    .image {
-      display: flex;
-      justify-content: center;
-      align-items: flex-start;
-
-      img {
-        transform: scaleX(-1);
-        height: 300px;
-        margin-left: 70px;
-      }
-
-      @media only screen and (max-width: 1100px) {
-        display: none;
-      }
-    }
+    padding-left: 100px;
+    padding-right: 100px;
   }
 
   .home-statistics {
@@ -257,29 +215,4 @@ const Filter = styled.div`
   }
 `;
 
-const StyledButton = withStyles({
-	root: {
-		background: 'linear-gradient(45deg, #21CBF3 30%, #48A7FF 90%)',
-		borderRadius: 2,
-		color: 'white',
-		width: 205,
-		padding: 0,
-		boxSizing: 'border-box',
-		boxShadow: '0 3px 5px 2px rgba(27, 68, 105, .3)',
-		fontWeight: 'bold',
-
-		'&:hover': {
-			background: 'linear-gradient(45deg, #21f39f 30%, #21CBF3 90%)',
-		},
-
-		'& span, a': {
-			width: '100%',
-			lineHeight: 2.5,
-		},
-	},
-	label: {
-		textTransform: 'capitalize',
-	},
-})(Button);
-
-export { HomeContainer, StyledButton, Side, Results, Practice, Filter };
+export { HomeContainer, Side, Results, Practice, Filter };
